% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_dropbox_revisions.R
\name{diff_dropbox_revisions}
\alias{diff_dropbox_revisions}
\title{Diff a CSV file's history}
\usage{
diff_dropbox_revisions(path, then, now = NULL)
}
\arguments{
\item{path}{relative to ~/Dropbox}

\item{then}{first datetime}

\item{now}{second datetime (defaults to \code{\link[=Sys.time]{Sys.time()}})}

\item{\dots}{path components (as in \code{\link[=file.path]{file.path()}})}
}
\description{
Compare two revisions of data stored in a .CSV file on Dropbox
}
\note{
DOES NOT YET WORK ON RSTUDIO SERVER. WILL SOMEDAY SOON.
}
\examples{
\dontrun{
library(inventory)

# Pick a file
csv_path <-
  my_dropbox("BY2015", "Throughputs",
             "BY2015-throughputs-area_sources.csv")

# How has this file changed in the last two weeks?
diff_dropbox_revisions(
  csv_path,
  now = Sys.time(),
  then = Sys.time() - ddays(14))

# You don't actually have to supply `now` ---
# but if you don't, you do have to say `then = `
diff_dropbox_revisions(
  csv_path,
  then = Sys.time() - ddays(14))

# You can diff two known revisions
# that differ by as little as a few mminutes
t1 <- "2017-09-08 09:10:00 PDT"
t2 <- "2017-09-08 10:10:00 PDT"
diff_dropbox_revisions(csv_path, t1, t2)

}

}
\seealso{
\code{\link[=ddays]{ddays()}}, \code{\link[=dhours]{dhours()}}, etc. (from the \code{lubridate} package)

\code{\link[=render_diff]{render_diff()}} and \code{\link[=diff_data]{diff_data()}}
}
