% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_exclude.R
\name{dropbox_exclude}
\alias{dropbox_exclude}
\title{Exclude file(s) or folder(s) from Dropbox}
\usage{
dropbox_exclude(...)
}
\arguments{
\item{\dots}{path components (as in \code{\link[=file.path]{file.path()}})}
}
\description{
Exclude file(s) or folder(s) from Dropbox
}
