% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dropbox_data.R
\name{get_dropbox_data}
\alias{get_dropbox_data}
\title{Get the version that existed at a certain time}
\usage{
get_dropbox_data(
  path,
  as_of = NULL,
  sheet = 1,
  ...,
  dtoken = NULL,
  verbose = getOption("verbose")
)
}
\arguments{
\item{path}{relative to ~/Dropbox}

\item{as_of}{most recent revision prior to or including this POSIXct}

\item{dtoken}{passed to \code{rdrop2} package methods}
}
\description{
Get the version that existed at a certain time
}
