% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dropbox_history.R
\name{get_dropbox_history}
\alias{get_dropbox_history}
\title{Get history from Dropbox API}
\usage{
get_dropbox_history(path, limit = 100)
}
\arguments{
\item{path}{relative to ~/Dropbox}

\item{dtoken}{passed to \code{rdrop2} package methods}

\item{rev_limit}{passed to \code{rdrop2} package methods}
}
\description{
Get history from Dropbox API
}
