% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_dropbox.R
\name{my_dropbox}
\alias{my_dropbox}
\title{Dropbox file/folder path(s)}
\usage{
my_dropbox(..., mustWork = FALSE, verbose = FALSE)
}
\arguments{
\item{\dots}{path components (as in \code{\link[=file.path]{file.path()}})}

\item{mustWork}{passed to normalizePath()}

\item{verbose}{logical}
}
\description{
Forms a path to file(s) or folders in your Dropbox.
Works just like \code{file.path("~", "Dropbox", ...)}
}
\note{
Looks first for "~/Dropbox (BAAQMD)". Failing that, looks for "~/Dropbox" instead. Throws an error if neither can be found.
}
\examples{
csv_path <-
  my_dropbox("BY2015", "Training", "ari", "02-input",
             "BY2015-throughputs-area_sources.csv")

file.exists(csv_path)

}
\seealso{
\code{\link[=file.path]{file.path()}}
}
