% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dropbox_timestamp.R
\name{parse_dropbox_timestamp}
\alias{parse_dropbox_timestamp}
\title{Parse timestamp(s) from the Dropbox API}
\usage{
parse_dropbox_timestamp(x, UTC = FALSE)
}
\arguments{
\item{x}{timestamps to be parsed (character)}

\item{UTC}{force UTC? (logical)}
}
\description{
Parse timestamp(s) from the Dropbox API
}
\note{
If \code{UTC} is \code{FALSE}, uses \code{Sys.timezone()} instead
}
