% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_dropbox.R
\name{path_dropbox}
\alias{path_dropbox}
\alias{my_dropbox}
\alias{dropbox_path}
\title{Dropbox file/folder path(s)}
\usage{
path_dropbox(..., mustWork = FALSE)

my_dropbox(..., mustWork = FALSE)

dropbox_path(..., mustWork = FALSE)
}
\arguments{
\item{\dots}{path components (as in \code{\link[=file.path]{file.path()}})}

\item{mustWork}{passed to \code{\link[=normalizePath]{normalizePath()}}}
}
\description{
Forms a path to file(s) or folders in your Dropbox.
Works just like \code{file.path("~", "Dropbox", ...)}
}
\section{Functions}{
\itemize{
\item \code{my_dropbox()}: alias

\item \code{dropbox_path()}: alias

}}
\note{
Looks for the following, in order:
\itemize{
\item "BAAQMD Dropbox"
\item "Dropbox (BAAQMD)"
\item "Dropbox"
}

Throws an error if none can be found.
}
\seealso{
\code{\link[=file.path]{file.path()}}
}
